Rename base package to reflect project name
