+++ src/main/java/com/github/fridujo/classpath/junit/extension/NoMatchingClasspathElementFoundException.java
@@ -1,0 +1,8 @@
+package com.github.fridujo.classpath.junit.extension;
+
+public class NoMatchingClasspathElementFoundException extends RuntimeException {
+
+    public NoMatchingClasspathElementFoundException(Gav gav) {
+        super(gav.toString() + " found no match in classpath");
+    }
+}
