+++ src/main/java/com/github/fridujo/classpath/junit/extension/buildtool/BuildTool.java
@@ -1,0 +1,13 @@
+package com.github.fridujo.classpath.junit.extension.buildtool;
+
+import java.util.Set;
+
+import com.github.fridujo.classpath.junit.extension.Gav;
+import com.github.fridujo.classpath.junit.extension.PathElement;
+
+public interface BuildTool {
+
+    Set<Artifact> listDependencies(PathElement path);
+
+    Set<Artifact> downloadDependency(Gav absoluteGav);
+}
