+++ src/main/java/com/github/fridujo/classpath/junit/extension/buildtool/NoBuildToolFoundException.java
@@ -1,0 +1,4 @@
+package com.github.fridujo.classpath.junit.extension.buildtool;
+
+public class NoBuildToolFoundException extends RuntimeException {
+}
