+++ src/main/java/com/github/fridujo/classpath/junit/extension/buildtool/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Currently supported Build Tools:
+ * <ul>
+ * <li>Maven</li>
+ * </ul>
+ */
+package com.github.fridujo.classpath.junit.extension.buildtool;
