--- src/main/java/com/github/fridujo/junit/extension/classpath/buildtool/NoBuildToolFoundException.java
@@ -1,4 +1,0 @@
-package com.github.fridujo.junit.extension.classpath.buildtool;
-
-public class NoBuildToolFoundException extends RuntimeException {
-}
