--- src/main/java/com/github/fridujo/junit/extension/classpath/buildtool/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Currently supported Build Tools:
- * <ul>
- * <li>Maven</li>
- * </ul>
- */
-package com.github.fridujo.junit.extension.classpath.buildtool;
