--- src/main/java/com/github/fridujo/junit/extension/classpath/jupiter/ModifiedClasspathExtension.java
@@ -1,67 +1,0 @@
-package com.github.fridujo.junit.extension.classpath.jupiter;
-
-import java.lang.reflect.Method;
-import java.util.Optional;
-
-import org.junit.jupiter.api.extension.ExtensionContext;
-import org.junit.jupiter.api.extension.InvocationInterceptor;
-import org.junit.jupiter.api.extension.ReflectiveInvocationContext;
-import org.junit.platform.commons.util.ReflectionUtils;
-
-import com.github.fridujo.junit.extension.classpath.Classpath;
-import com.github.fridujo.junit.extension.classpath.buildtool.BuildTool;
-
-class ModifiedClasspathExtension implements InvocationInterceptor {
-
-    private final ExtensionContext.Namespace namespace = ExtensionContext.Namespace.create(ModifiedClasspathExtension.class);
-
-    public void interceptTestMethod(Invocation<Void> invocation,
-                                    ReflectiveInvocationContext<Method> invocationContext,
-                                    ExtensionContext extensionContext) {
-        invocation.skip();
-
-        ExtensionContext.Store store = extensionContext.getRoot().getStore(namespace);
-        BuildTool buildTool = store.get(BuildTool.class, BuildTool.class);
-
-        ModifiedClasspath annotation = invocationContext.getExecutable().getAnnotation(ModifiedClasspath.class);
-        Classpath classpath = Classpath.current(buildTool)
-            .removeJars(annotation.excludeJars())
-            .removeDependencies(annotation.excludeDependencies())
-            .addDependencies(annotation.addDependencies());
-        store.put(BuildTool.class, classpath.buildTool);
-        ClassLoader modifiedClassLoader = classpath
-            .newClassLoader();
-
-        ClassLoader currentThreadPreviousClassLoader = replaceCurrentThreadClassLoader(modifiedClassLoader);
-
-        try {
-            invokeMethodWithModifiedClasspath(
-                invocationContext.getExecutable().getDeclaringClass().getName(),
-                invocationContext.getExecutable().getName(),
-                modifiedClassLoader);
-        } finally {
-            Thread.currentThread().setContextClassLoader(currentThreadPreviousClassLoader);
-        }
-    }
-
-    private ClassLoader replaceCurrentThreadClassLoader(ClassLoader modifiedClassLoader) {
-        ClassLoader currentThreadPreviousClassLoader = Thread.currentThread().getContextClassLoader();
-        Thread.currentThread().setContextClassLoader(modifiedClassLoader);
-        return currentThreadPreviousClassLoader;
-    }
-
-    private void invokeMethodWithModifiedClasspath(String className, String methodName, ClassLoader classLoader) {
-        final Class<?> testClass;
-        try {
-            testClass = classLoader.loadClass(className);
-        } catch (ClassNotFoundException e) {
-            throw new IllegalStateException("Cannot load test class [" + className + "] from modified classloader, verify that you did not exclude a path containing the test", e);
-        }
-
-        Object testInstance = ReflectionUtils.newInstance(testClass);
-        final Optional<Method> method = ReflectionUtils.findMethod(testClass, methodName);
-        ReflectionUtils.invokeMethod(
-            method.orElseThrow(() -> new IllegalStateException("No test method named " + methodName)),
-            testInstance);
-    }
-}
