--- src/main/java/com/github/fridujo/junit/extension/classpath/utils/SystemVariables.java
@@ -1,10 +1,0 @@
-package com.github.fridujo.junit.extension.classpath.utils;
-
-import java.util.Optional;
-
-public class SystemVariables {
-
-    public String get(String name) {
-        return Optional.ofNullable(System.getProperty(name)).orElse(System.getenv(name));
-    }
-}
