+++ src/main/resources/META-INF/services/com.github.fridujo.classpath.junit.extension.buildtool.BuildToolResolver
@@ -1,0 +1,1 @@
+com.github.fridujo.classpath.junit.extension.buildtool.maven.MavenResolver
