--- src/main/resources/META-INF/services/com.github.fridujo.junit.extension.classpath.buildtool.BuildToolResolver
@@ -1,1 +1,0 @@
-com.github.fridujo.junit.extension.classpath.buildtool.maven.MavenResolver
